Move helpers folder into utils
