+++ Source/SelfService/Web/utils/helpers/dates.ts
@@ -69,4 +69,3 @@ export const formatStartingDate = (initialDate: string) => {
 export const formatDate = (date: Date): string => {
     return format(date, 'yyyy/MM/dd');
 };
-
