+++ Source/SelfService/Web/utils/helpers/downloads.tsx
@@ -5,7 +5,7 @@ import React, { useEffect, useState } from 'react';
 
 import { useSnackbar } from 'notistack';
 
-import { getPodLogs } from '../../../apis/solutions/api';
+import { getPodLogs } from '../../apis/solutions/api';
 
 import { GridRenderCellParams } from '@mui/x-data-grid-pro';
