+++ Source/SelfService/Web/utils/helpers/index.ts
@@ -1,8 +1,15 @@
 // Copyright (c) Dolittle. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-export { capitalize, removeStringPattern, removeStringPatternAndCapitalize, getRuntimeNumberFromString } from './strings';
-export { formatTime, formatStartingDate } from '../../../utils/dates';
+export {
+    capitalize,
+    removeStringPattern,
+    removeStringPatternAndCapitalize,
+    getRuntimeNumberFromString,
+    trimPrefix,
+    trimSuffix,
+} from './strings';
+export { formatTime, formatStartingDate } from './dates';
 export { DownloadLogs } from './downloads';
 export { formatBytes } from './numbers';
 export { isAlphaNumeric } from './regex';
