+++ Source/SelfService/Web/utils/helpers/regex.ts
