+++ Source/SelfService/Web/utils/helpers/strings.ts
@@ -20,3 +20,13 @@ export const getRuntimeNumberFromString = (s: string) => {
     if (typeof s !== 'string') return 'N/A';
     return removeStringPatternAndCapitalize(s, /dolittle\/runtime:/gi);
 };
+
+export function trimPrefix(str: string, prefix: string): string {
+    if (str.startsWith(prefix)) return str.slice(prefix.length);
+    else return str;
+};
+
+export function trimSuffix(str: string, suffix: string): string {
+    if (str.endsWith(suffix)) return str.slice(0, -(suffix.length));
+    else return str;
+};
